Funci'on tick en tick

+++ Modelo/src/modelo/Bebida.java
@@ -14,7 +14,7 @@ public class Bebida extends Producto{
     public String tipo;     //frio o caliente  
     public int tamano;      //200ml, 250ml o 330ml
 
-    public Bebida(String nombre, int precio, double tiempoProduccion, String tipo, int tamano) {
+    public Bebida(String nombre, int precio, int tiempoProduccion, String tipo, int tamano) {
         super(nombre, precio, tiempoProduccion);
         this.tipo = tipo;
         this.tamano = tamano;
